Move DesignerActionToolStripDropDown to its own file and make DesignerActionUI a partial class
